added 3 colors to the gradient

+++ pydiffvg/save_svg.py
@@ -145,7 +145,7 @@ def save_svg(filename, width, height, shapes, shape_groups, use_gamma = False):
             else:
                 c = shape_group.stroke_color.data.cpu().numpy()
                 shape_node.set('stroke', 'rgb({}, {}, {})'.format(\
-                    int(255 * c[0]), int(255 * c[1]), int(255 * c[2])))
+                    255 * c[0], 255 * c[1], 255 * c[2]))
                 shape_node.set('stroke-opacity', str(c[3]))
             shape_node.set('stroke-linecap', 'round')
             shape_node.set('stroke-linejoin', 'round')
